Fix startUnclean option being ignored
Problem: I don't know if this is new behavior, but if you pass `temp` to
SSB-DB it will `rm -rf` the directory no matter what. This means that if
you pass `{ startUnclean: true }` to Scuttle-Testbot that it will be
ignored. Separation of concerns is hard when you've got lots of modules
stacked on top of each other with similar behavior!

Solution: Don't pass `temp`, don't trust SSB-DB, and just come up with
the path to the temporary directory by ourselves.